fix: separacion fichero de clases
